Move UserDataUpdateEventType to separate file
